+++ packages/perseus/src/widgets/interactive-graph.tsx
@@ -1811,11 +1811,7 @@ class InteractiveGraph extends React.Component<Props, State> {
             return (
                 <StatefulMafsGraph
                     {...this.props}
-                    showLabelsFlag={
-                        this.props.apiOptions?.flags?.["mafs"]?.[
-                            "interactive-graph-locked-features-labels"
-                        ]
-                    }
+                    flags={this.props.apiOptions?.flags}
                     ref={this.mafsRef}
                     gridStep={gridStep}
                     snapStep={snapStep}
@@ -2697,7 +2693,7 @@ const staticTransform = _.identity;
 
 export default {
     name: "interactive-graph",
-    displayName: "Interactive graph",
+    displayName: "Interactive graph (Assessments only)",
     widget: InteractiveGraph,
     staticTransform: staticTransform,
 } as WidgetExports<typeof InteractiveGraph>;