+++ packages/perseus/src/widgets/measurer/measurer-validator.ts
@@ -1,0 +1,10 @@
+function measurerValidator() {
+    return {
+        type: "points",
+        earned: 1,
+        total: 1,
+        message: null,
+    };
+}
+
+export default measurerValidator;